Add includeZones support to ATW devices.
Using an Ecodan ATW heat pump with FTC6 controller, when adding the heat pump as a 'thermostat', three Home Accessories are created. One is Zone1, which is not optional, but I do not use. One is unknown, presumed Zone2, although this is inactive on my FTC6, MELCloud app, and is already catered for in the config schema. The third is the Hot Water cylinder, which I wanted to keep.
This feature adds a config setting to check/uncheck Zone1/Zone2/Hot Water.
This allows me to see only the useful thermostat in my Apple Home app.
Not an excellent coder, this was largely created with AI support.